* arts-2-1-611
* src/matpack*.{h,cc}:

(Sparse):
Don't use pointer variables for the members mdata, mrowind and
mcolptr.
Several advantages: No custom memory management needed -> Copy
constructor, copy assignment operator and destructor removed. Code is
much more readable.
(mult, transpose): Non-const iterators were used to loop over elements
of the const input parameter.

(*Iterator*):
Remove copy constructor from all Iterator classes. The implicit one
presence of a custom copy constructor will prevent the generation of
an implicit move constructor. Might give better performance.

(Vector, Matrix, TensorX):
Change the argument type of the assignment operator from a const
reference to a call by value. This way, we can let the copy
constructor do the work instead of duplicating its initialization code
in the assignment op. Additionally, with this implementation, the copy
assignment operator will also work as a move assignment operator in
C++11. And this implementation is exception safe. With the previous
one, if the copy operation failed, the current object was left in an
undetermined state.
Add swap function to swap two objects.

* src/xml_io_basic_types.cc: Adapt for Sparse changes.

* src/matpack.h: Include debug.h

* src/debug.h: Include iostream, add some std::.

* controlfiles/instruments/odinsmr/yREFERENCE*.xml: Add ref values.

* controlfiles/instruments/odinsmr/TestOdinSMR*.arts:

Compare y against reference.

* src/m_fos.cc (fos):

Add explicit cast to Numeric for 'nin' in division to silence compiler
warning.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@8391 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/m_fos.cc
@@ -574,7 +574,7 @@ void fos(
                               ssource0(iv,joker) += sp;
                             }
                         }
-                      ssource0 *= 4*PI/nin;
+                      ssource0 *= 4*PI/(Numeric)nin;
                     }
                  
                   // RT of ppath step
